[fix] colorizer-64x64 layer filter added to discriminator [refactor] sampler moved to basegan
