+++ examples/common.py
@@ -371,9 +371,6 @@ class TextInput:
             return string
 
 
-def lookup_sampler(name):
-    return CLI.sampler_for(name, name)
-
 def parse_size(size):
     width = int(size.split("x")[0])
     height = int(size.split("x")[1])
